+++ Coditech.Project/API/Comman/Coditech.API.Client/Endpoint/HospitalDoctors/HospitalDoctorsEndpoint.cs
@@ -13,7 +13,7 @@ namespace Coditech.API.Endpoint
         }
         public string CreateHospitalDoctorsAsync() =>
             $"{CoditechAdminSettings.CoditechHospitalManagementSystemApiRootUri}/HospitalDoctors/CreateHospitalDoctors";
-        public string GetHospitalDoctorsAsync(long hospitalDoctorId) =>
+        public string GetHospitalDoctorsAsync(int hospitalDoctorId) =>
             $"{CoditechAdminSettings.CoditechHospitalManagementSystemApiRootUri}/HospitalDoctors/GetHospitalDoctors?hospitalDoctorId={hospitalDoctorId}";
 
         public string UpdateHospitalDoctorsAsync() =>
